Add version/run commands
The main purpose of this was to add a version command. At the same time,
it was necessary to make a run command to differentiate a 'version'
command sent to hosts from the version subcommand.

This could have been achieved using a --version (-V) commandline flag;
however, the golang CLI standard seems to be a version subcommand and
not a flag. Additionally, copy (scp) functionality is planned for the
future, which will require an additional command, so this refactor will
be useful later.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>