refactor tentacle to config w/ result
Tentacle was returning result information from hosts, but to be
extensible (planning for copy functionality), it is better to have a
struct for common configs (Tentacle) and a different one for results
(Results). Tentacle is now an interface that defines the "Do" method,
which tentacles may define to perform an action. Command is now a
command tentacle capable of running commands, and copy will later be a
different type of tentacle.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>