copy: fix bug where files weren't copied
Instead of sftp.File.WriteTo(local os.File) to write from a local file to a
remote one, it should be sftp.File.ReadFrom(local os.File). What is
unexpected is that WriteTo was not producing any errors trying to write
to a readonly local file. Integration/functional tests should help with
this issue in the future.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>