Do not use custom-written group variable parser
Instead of writing a custom parser in Go, rely on Bash to parse
variables automatically. Use `env` and `bash` commands to achieve the
functionality. The downside of this is that variables in the host groups
file must now be `export`-ed to be used by Octopus.

In some ways, this is a feature because `export`-ed variables were not
previously supported. And forcing variables to use `export` isn't a bad
requirement, as there is a shellcheck linter warning if unused variables
aren't exported, and the host groups file will have mostly unused
variables.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>