Fix minor bug in --host-groups flag completion
The previous completion would suggest completions given host groups in a
file when `<tab>``<tab>`-ing after the `--host-groups` or `-g` flag, but
if the user started typing the name of a group, the suggestions would
not continue to be filled in.

Take the overide flags behavior directly from kubectl's custom command
definition, and use that since it implements the behavior already. This
function is probably a bit overkill for Octopus' simple needs, but it
works.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>